+++ src/main/java/io/github/astrapi69/file/FileExtensions.java
@@ -32,19 +32,15 @@ import java.nio.channels.FileLock;
 import java.nio.file.Files;
 
 import io.github.astrapi69.file.read.ReadFileExtensions;
-import io.github.astrapi69.io.file.FilenameExtensions;
 
 /**
- * Utility class for the use of File object.
+ * Utility class for the use with File object
  *
  * @version 1.0
  * @author Asterios Raptis
  */
 public final class FileExtensions
 {
-
-	/** The Constant VELOCITY_TEMPLATE_FILE_EXTENSION. */
-	public static final String VELOCITY_TEMPLATE_FILE_EXTENSION = ".vm";
 
 	private FileExtensions()
 	{
@@ -96,35 +92,6 @@ public final class FileExtensions
 	}
 
 	/**
-	 * Gets the filename with the absolute path prefix.
-	 *
-	 * @param file
-	 *            the file.
-	 * @return the filename prefix
-	 * @deprecated use instead the same name method in class {@link FilenameExtensions} from module
-	 *             silly-io. Note that this method will be deleted in next major release
-	 */
-	public static String getFilenamePrefix(final File file)
-	{
-		return FilenameExtensions.getFilenamePrefix(file);
-	}
-
-	/**
-	 * Gets the filename suffix or null if no suffix exists or the given file object is a directory.
-	 *
-	 * @param file
-	 *            the file.
-	 * @return 's the filename suffix or null if no suffix exists or the given file object is a
-	 *         directory.
-	 * @deprecated use instead the same name method in class {@link FilenameExtensions} from module
-	 *             silly-io. Note that this method will be deleted in next major release
-	 */
-	public static String getFilenameSuffix(final File file)
-	{
-		return FilenameExtensions.getFilenameSuffix(file);
-	}
-
-	/**
 	 * Gets the content type from the given file object
 	 *
 	 * @param file
@@ -139,20 +106,6 @@ public final class FileExtensions
 	}
 
 	/**
-	 * Gets the filename without the extension or null if the given file object is a directory.
-	 *
-	 * @param file
-	 *            the file.
-	 * @return the filename without the extension or null if the given file object is a directory
-	 * @deprecated use instead the same name method in class {@link FilenameExtensions} from module
-	 *             silly-io. Note that this method will be deleted in next major release
-	 */
-	public static String getFilenameWithoutExtension(final File file)
-	{
-		return FilenameExtensions.getFilenameWithoutExtension(file);
-	}
-
-	/**
 	 * Not yet implemented. Checks if the given file is open.
 	 *
 	 * @param file
@@ -164,10 +117,9 @@ public final class FileExtensions
 	public static boolean isOpen(final File file) throws IOException
 	{
 		boolean open = false;
-		FileLock lock = null;
 		try (RandomAccessFile fileAccess = new RandomAccessFile(file.getAbsolutePath(), "rw"))
 		{
-			lock = fileAccess.getChannel().tryLock();
+			FileLock lock = fileAccess.getChannel().tryLock();
 			if (lock == null)
 			{
 				open = true;