+++ src/main/java/module-info.java
@@ -1,0 +1,11 @@
+module file.worker {
+    requires silly.io.main;
+    requires throwable;
+    requires org.apache.commons.io;
+    requires org.apache.commons.lang3;
+    requires checksum.up;
+    requires crypt.api;
+    requires java.logging;
+    requires silly.collection.main;
+    requires silly.strings;
+}